DownwardMetrics: Add new API field for virtio serial and admitter
It adds a new API spec for DownwardMetrics virtio serial support.
The following spec needs to be added to expose the DownwardMetrics
with virtio serial to the vm:

```
spec:
  domain:
    devices:
      downwardMetrics: {}
```

Also, it removes the need to add a disk and volume for the
downwardMetric since now the downwardMetrics can be exposed through
virtio serial.
Futhermore, it adds a check to catch invalid DownwardMetrics
configurations, i.e., virtio serial enabled without the DownwardMetrics
featureGate set.

Signed-off-by: German Maglione <gmaglione@redhat.com>
Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>